requirements.txt
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
streamlit==1.15.1
